# Rewrite of the inference system