--- a/myia/abstract/__init__.py
+++ b/myia/abstract/__init__.py
@@ -12,8 +12,8 @@ from .base import (  # noqa
 
 from .inf import (  # noqa
     AbstractTrack,
-    XInferrer,
-    GraphXInferrer,
+    Inferrer,
+    GraphInferrer,
     from_value,
     to_abstract,
 )