--- a/myia/abstract/__init__.py
+++ b/myia/abstract/__init__.py
@@ -1,4 +1,34 @@
 
+from .core import (  # noqa
+    InferenceLoop,
+    EvaluationCache,
+    find_coherent_result,
+    reify,
+)
+
+from .graph_infer import (  # noqa
+    Context,
+    Contextless,
+    CONTEXTLESS,
+    Track,
+    InferenceEngine,
+    AbstractReference,
+    Reference,
+    VirtualReference,
+)
+
+from .utils import (  # noqa
+    ANYTHING,
+    InferenceError,
+    MyiaTypeError,
+    MyiaShapeError,
+    Unspecializable,
+    VOID,
+    DEAD,
+    POLY,
+    INACCESSIBLE,
+)
+
 from .base import (  # noqa
     ABSENT,
     AbstractBase,