--- a/myia/abstract/__init__.py
+++ b/myia/abstract/__init__.py
@@ -1,5 +1,5 @@
 
-from .core import (  # noqa
+from .loop import (  # noqa
     InferenceLoop,
     InferenceTask,
     Pending,