--- a/myia/abstract/__init__.py
+++ b/myia/abstract/__init__.py
@@ -49,25 +49,23 @@ from .data import (  # noqa
 )
 
 
-from .graph_infer import (  # noqa
+from .ref import (  # noqa
     Context,
     Contextless,
     CONTEXTLESS,
-    Track,
     AbstractReference,
     Reference,
     VirtualReference,
-    InferenceEngine,
     EvaluationCache,
 )
 
 
 from .inf import (  # noqa
-    AbstractTrack,
     Inferrer,
     GraphInferrer,
     from_value,
     to_abstract,
+    InferenceEngine,
 )