--- a/myia/abstract/__init__.py
+++ b/myia/abstract/__init__.py
@@ -10,7 +10,8 @@ from .loop import (  # noqa
 )
 
 from .utils import (  # noqa
-    to_value,
+    build_value,
+    build_type,
     abstract_clone,
     abstract_clone_async,
     concretize_abstract,