--- a/myia/abstract/__init__.py
+++ b/myia/abstract/__init__.py
@@ -1,7 +1,6 @@
 
 from .loop import (  # noqa
     InferenceLoop,
-    InferenceTask,
     Pending,
     PendingFromList,
     PendingTentative,