--- a/myia/abstract/__init__.py
+++ b/myia/abstract/__init__.py
@@ -51,7 +51,7 @@ from .data import (  # noqa
     AbstractClass,
     AbstractJTagged,
     TrackDict,
-    Subtrack,
+    Track,
     VALUE,
     TYPE,
     SHAPE,