--- a/myia/abstract/__init__.py
+++ b/myia/abstract/__init__.py
@@ -29,7 +29,6 @@ from .data import (  # noqa
     VOID,
     DEAD,
     POLY,
-    NOTVISIBLE,
     Possibilities,
     PrimitiveFunction,
     GraphFunction,