--- a/myia/abstract/base.py
+++ b/myia/abstract/base.py
@@ -12,7 +12,7 @@ from ..prim import Primitive
 from ..debug.utils import mixin
 from ..utils import overload, UNKNOWN, Named
 
-from .core import Pending, is_simple, PendingTentative
+from .loop import Pending, is_simple, PendingTentative
 from .graph_infer import Reference, Context
 
 from .data import (  # noqa