This PR changes the inference system to be overall simpler and easier to work with.

Changes to the API:

* The inferrer now lives in `myia.abstract`
* Instead of having value, type, shape tracks, the inferrer works with instances of `AbstractValue` and its subclasses, e.g. `AbstractTuple`. An abstract value can contain elements and it can also have a value, a type or a shape, in its `values` dictionary. At the moment, only `AbstractArray` has a shape. Not all abstract values have a value or a type.
* The argspec for the pipeline must now be a tuple of AbstractValues. A Python object can be converted to an AbstractValue with `myia.abstract.from_value(x, broaden=True/False)`.
* `node.inferred` is gone, replaced by `node.abstract`.
* AbstractValues don't keep a reference to the engine, unlike Inferrers in the previous system.
* There are still Inferrer objects but they are kept in the engine.
* The engine now tracks the set of functions that can be called at any given point in the program.
* NOSHAPE, TupleShape, ListShape, etc. are all gone.
* The old equivalence checker is replaced by `myia.abstract.utils.amerge`.
* I had to move pipeline creation back where it was in the `@myia` decorator because it didn't work otherwise.

Changes to tests:

* The tests in `test_infer` now tests all tracks at once. About 50 tests were cut overall.
* The tests run about twice as fast (47 seconds vs 92 seconds for the current master on my machine). I am not certain where all of the difference comes from, though I know one particular improvement accounts for about 15 seconds, and maybe 4-5 seconds worth of cut tests. Either way, it's a welcome improvement.