@abergeron 2019-02-08
This results in a speedup of about 50% on the model specialize_backwards test (from ~6s to ~4s).

However the measurement is hard since the compile phase is having mood swings ranging from 1s to 2.5s.  I'll check what's up with that later.